Game 01 | Shadow Clone | Added Clone entity, and ClonedPlayer entity
Added the new Clone and ClonedPlayer entities.
The Clone is another abstract entity (maybe some monsters will inherit from it), that receives a position each frame and uses it.
The cloned player also receives additional info for animations.

The player also received a new component named PositionBuffer, that stores its position each frame, and updates it but with a time delay.