Game 01 | Shadow Clone | Trampoline Jump for Player
Added the ability for the player to perform a trampoline jump, by jumping when touching the clone and in the air. This jump reaches super high, and restores all the player's normal jumps, allowing them to even chain multiple trampoline jumps if they are skilled enough.

This feature was designed using the component pattern, always keeping in mind how different entities react to each other.